gatepassr_api/requirements.txt
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
Flask==3.0.2
Flask-Session==0.6.0
gunicorn==21.2.0
identity==0.5.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msal==1.26.0
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.1
